requirements.txt
requests
beautifulsoup4
openai
python-dotenv
praw
